requirements.txt
Django==3.1.6
django-allauth==0.45.0
django-environ==0.4.5
gunicorn
django-phonenumber-field[phonenumbers]
Babel==2.9.1
django-extensions==3.1.3
python-binance==1.0.12